requirements.txt
requests
beautifulsoup4
playwright
supabase
pandas
python-dotenv
lxml
